refactor: decompose utils.py into domain-specific submodules
Decompose the monolithic utils.py (878 lines) into focused modules:

- pdfminer/geometry/matrix.py — 2D affine matrix operations
  (parse_rect, mult_matrix, translate_matrix, apply_matrix_pt,
   apply_matrix_rect, apply_matrix_norm, MATRIX_IDENTITY)

- pdfminer/geometry/spatial.py — Bounding boxes, distances, Plane index
  (get_bound, vecBetweenBoxes, Plane)

- pdfminer/prediction.py — PNG/TIFF predictor filter reversal
  (paeth_predictor, apply_png_predictor, apply_tiff_predictor)

- pdfminer/textencoding.py — PDF text encoding/decoding, AES unpadding
  (decode_text, unpad_aes, PDFDocEncoding)

- pdfminer/formatting.py — Number and string formatting
  (format_int_roman, format_int_alpha, bbox2str, matrix2str,
   enc, encode_xml_safe)

utils.py is now a thin re-export layer maintaining full backward
compatibility — no other files need to change.

Naming improvements:
- Added encode_xml_safe() as a clear alias for enc()
- Added comprehensive docstrings to every extracted function
- Each module has a clear domain and single responsibility

Refactoring verified with 4-level safety check:
1. All 14 Regrets clusters GREEN after refactor
2. Direct output matches pre-refactor ground truth (48 outputs)
3. Fingerprint cross-match with KEBENARAN 2 (14 fingerprints)
4. Chain hash validation (3 chains all match)
